requirements.txt
openai
python-dotenv
tushare
akshare
pandas
yfinance
matplotlib
pypdf
chromadb
sentence-transformers
